Resolution and number of SDFs can be adjusted
Going into the SDFTextureCollection scriptable object you can now find parameters for the count of SDFs in the array and the resolution of each SDF. This enables use cases like:

1 very high quality (512x512x512) SDF
8 high quality (256x256x256) SDF
64 medium quality (128x128x128) SDF
etc

You can now place any resolution of SDF into the mix and it should attempt to scale. I've noticed some artifacting in the process.