access statics from SDFTextureCollection
First steps to making the SDFTextureCollection accessible in Jobs. Currently there's a memory leak because I have to remake the NativeArray every time I want to update it. Committing this mostly to put a pin in the idea.

Currently the idea is to dump all the texture data into one big array and then index that carefully in jobs system. Hopefully I can work back to having native arrays for each texture.